Move types to types.py
