+++ examples/customize/build_graph/pipeline/pipeline_with_notifications.py
@@ -14,8 +14,8 @@ from neo4j_graphrag.experimental.components.text_splitters.fixed_size_splitter i
     FixedSizeSplitter,
 )
 from neo4j_graphrag.experimental.pipeline import Pipeline
-from neo4j_graphrag.experimental.pipeline.notification import Event
 from neo4j_graphrag.experimental.pipeline.pipeline import PipelineResult
+from neo4j_graphrag.experimental.pipeline.types import Event
 
 logger = logging.getLogger(__name__)
 logging.basicConfig()
