+++ src/neo4j_graphrag/experimental/pipeline/types.py
@@ -17,7 +17,8 @@ from __future__ import annotations
 import datetime
 import enum
 from collections import defaultdict
-from typing import Any, Optional, Union
+from collections.abc import Awaitable
+from typing import Any, Optional, Protocol, Union
 
 from pydantic import BaseModel, ConfigDict, Field
 
@@ -60,6 +61,45 @@ class RunResult(BaseModel):
     timestamp: datetime.datetime = Field(
         default_factory=lambda: datetime.datetime.now(datetime.timezone.utc)
     )
+
+
+class EventType(enum.Enum):
+    PIPELINE_STARTED = "PIPELINE_STARTED"
+    TASK_STARTED = "TASK_STARTED"
+    TASK_FINISHED = "TASK_FINISHED"
+    PIPELINE_FINISHED = "PIPELINE_FINISHED"
+
+    @property
+    def is_pipeline_event(self) -> bool:
+        return self in [EventType.PIPELINE_STARTED, EventType.PIPELINE_FINISHED]
+
+    @property
+    def is_task_event(self) -> bool:
+        return self in [EventType.TASK_STARTED, EventType.TASK_FINISHED]
+
+
+class Event(BaseModel):
+    event_type: EventType
+    run_id: str
+    """Pipeline unique run_id, same as the one returned in PipelineResult after pipeline.run"""
+    timestamp: datetime.datetime
+    message: Optional[str] = None
+    """Optional information about the status"""
+    payload: Optional[dict[str, Any]] = None
+    """Input or output data depending on the type of event"""
+
+
+class PipelineEvent(Event):
+    pass
+
+
+class TaskEvent(Event):
+    task_name: str
+    """Name of the task as defined in pipeline.add_component"""
+
+
+class EventCallbackProtocol(Protocol):
+    def __call__(self, event: Event) -> Awaitable[None]: ...
 
 
 EntityInputType = Union[str, dict[str, Union[str, list[dict[str, str]]]]]
